# poslappapi/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /roles:
    x-swagger-router-controller: roles.controller
    get:
      description: get the role name list
      operationId: getRoles
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRolesListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: post the role name in list
      operationId: addRol
      parameters:
        - in: body
          name: nombreRol
          description: The Rol to create.
          schema:
            $ref: "#/definitions/Roles"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetRolesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /roles/{id}:
    x-swagger-router-controller: roles.controller
    get:
      description: get the role name list
      operationId: getRolesbyPk
      parameters:
        - name: id
          description: rol id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRolesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a game system
      operationId: updateRoles
      # define the parameters
      parameters:
        - name: id
          description: game system id
          type: number
          in: path
          required: true
        - in: body
          name: nombreRol
          description: The Game System to update.
          schema:
            $ref: "#/definitions/Roles"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetRolesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    delete:
      description: delete a rol
      # define the parameters
      operationId: deleteRoles
      parameters:
        - name: id
          description: rol id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Roles:
    type: object
    properties:
      # id_rol:
      #   type: integer
      #   description: Name of the Game System
      nombreRol:
        type: string
        description: Description of name rol
    required:
      #- id_rol
      # - id
      - nombreRol
  GetRolesResponse:
    type: object
    properties:
      # id_rol:
      #   type: integer
      #   description: identifier
      id:
        type: number
        description: Identifier
      nombreRol:
        type: string
        description: Name of the Game System
    required:
      # - id_rol
      # - id
      - nombreRol
  GetRolesListResponse:
    required:
      - rol
    properties:
      rol:
        type: array
        items:
          $ref: "#/definitions/GetRolesResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string